Refactors code

+++ test/params_ex_tests.cpp
@@ -35,7 +35,7 @@ namespace
         return nullptr;
         }
 
-    int bad_fread( [[maybe_unused]] void* ptr,
+    size_t bad_fread( [[maybe_unused]] void* ptr,
         [[maybe_unused]] size_t size,
         [[maybe_unused]] size_t count,
         [[maybe_unused]] FILE* stream )
